pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "deep_learning"
version = "0.1.0"
description = "Deep Learning framework"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.19.0",
    "pandas>=1.1.0",
    "scikit-learn>=0.24.0",
    "matplotlib>=3.3.0",
    "seaborn>=0.11.0",
    "pyyaml>=5.4.0",
    "joblib>=1.0.0",
    "mlflow>=1.12.0",
    "optuna>=2.3.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 119
target-version = ['py311']
